+++ CountriesOfTheWorld/CountriesOfTheWorldTests/CountriesControllerIPhoneUnitTests.swift
@@ -1,5 +1,5 @@
 //
-//  CountriesControllerUnitTests.swift
+//  CountriesControllerIPhoneUnitTests.swift
 //  CountriesOfTheWorldTests
 //
 //  Created by Volodymyr Seredovych on 30.12.2021.
@@ -108,36 +108,5 @@ class CountriesControllerIPhoneUnitTests: XCTestCase {
                       textDidChange: "")
         let isFilterCountriesEmpty = try XCTUnwrap(sut.filteredCountries?.isEmpty)
         XCTAssertTrue(isFilterCountriesEmpty)
-    }
-}
-
-class CountriesControllerIPadUnitTests: XCTestCase {
-    var sut: CountriesController!
-    var window: UIWindow!
-    var networkManager = NetworkManager(client: ApolloMock())
-
-    override func setUp() {
-        window = UIWindow()
-        sut = CountriesController(networkManager, .iPad)
-        window.rootViewController = sut
-        super.setUp()
-    }
-
-    override func tearDown() {
-        window = nil
-        sut = nil
-        super.tearDown()
-    }
-
-    func testHeaghtForHeaderInSectionForIPad() {
-        let expectedHeight: CGFloat = 0
-        let actualHeight = sut.tableView(sut.tableView, heightForHeaderInSection: 0)
-        XCTAssertEqual(expectedHeight, actualHeight)
-    }
-
-    func testTableIPadViewHeaderType() {
-        let header = sut.tableView(sut.tableView, viewForHeaderInSection: 0)
-        let isCurrentType = header != nil
-        XCTAssertTrue(isCurrentType)
     }
 }
