+++ CountriesOfTheWorld/CountriesOfTheWorldTests/CountryDetailsControllerIPhoneUnitTests.swift
@@ -1,5 +1,5 @@
 //
-//  CountriesDetailsControllerUnitTests.swift
+//  CountryDetailsControllerIPhoneUnitTests.swift
 //  CountriesOfTheWorldTests
 //
 //  Created by Volodymyr Seredovych on 03.01.2022.
@@ -63,42 +63,5 @@ class CountryDetailsControllerIPhoneUnitTests: XCTestCase {
         sut.addLabelsToStackView()
         XCTAssertTrue(sut.stackView.subviews.allSatisfy({ $0 is DetailsFieldPlaceHolderView }))
         XCTAssertTrue(sut.stackView.subviews.count == 6)
-    }
-}
-
-class CountriesDetailsControllerIPadUnitTests: XCTestCase {
-
-    typealias Continent = CountriesQuery.Data.Country.Continent
-    var sut: CountryDetailsController!
-    var networkManager = NetworkManager(client: ApolloMock())
-    var window: UIWindow!
-
-    override func setUp() {
-        window = UIWindow()
-        sut = CountryDetailsController(networkManager, .iPad)
-        window.rootViewController = sut
-        super.setUp()
-    }
-
-    override func tearDown() {
-        window = nil
-        sut = nil
-        super.tearDown()
-    }
-
-    func testiPadControllerSubviewsOrderWithoutData() {
-        sut.fillDetailsViewWithCountryQuery()
-        sut.viewDidLoad()
-        XCTAssertTrue(sut.view.subviews.count == 1)
-    }
-
-    func testIPadControllerSubviewsOrderWithData() {
-        sut.loadData(code: "UA")
-        sut.fillDetailsViewWithCountryQuery()
-        sut.viewDidLoad()
-        XCTAssertTrue(sut.view.subviews[0] is CountriesDetailsHeader)
-        XCTAssertTrue(sut.view.subviews[1] is UIScrollView)
-        XCTAssertTrue(sut.view.subviews[1].subviews[1] is UIImageView)
-        XCTAssertTrue(sut.view.subviews[1].subviews[2] is UIStackView)
     }
 }
